[Rename] from `editorCommands.ts` to `command.contrib.ts`
